rename and disable hooks
